changed FloatField for PositiveIntegerField

+++ course/models.py
@@ -46,7 +46,7 @@ class Payment(models.Model):
     pay_day = models.DateTimeField(auto_now_add=True, verbose_name='Дата оплаты')
     paid_course = models.ForeignKey(Course, on_delete=models.SET_NULL, **NULLABLE, related_name='payments', verbose_name='Оплаченный курс')
     paid_lesson = models.ForeignKey(Lesson, on_delete=models.SET_NULL, **NULLABLE, related_name='payments', verbose_name='Оплаченный урок')
-    paid_amt = models.FloatField(verbose_name='Сумма оплаты')
+    paid_amt = models.PositiveIntegerField(verbose_name='Сумма оплаты')
     payment_option = models.CharField(max_length=4, choices=PAYMENT_CHOICES, verbose_name='Способ оплаты')
